# Add various performance improvements