--- a/src/array.ts
+++ b/src/array.ts
@@ -1,7 +1,7 @@
 import { isStateTreeNode, types } from "mobx-state-tree";
-import { BaseType, setEnv, setParent } from "./base";
+import { BaseType } from "./base";
 import { ensureRegistered } from "./class-model";
-import { $readOnly, $type } from "./symbols";
+import { $env, $parent, $readOnly, $type } from "./symbols";
 import type { IAnyStateTreeNode, IAnyType, IArrayType, IMSTArray, IStateTreeNode, Instance, InstantiateContext } from "./types";
 
 export class QuickArray<T extends IAnyType> extends Array<Instance<T>> implements IMSTArray<T> {
@@ -9,7 +9,16 @@ export class QuickArray<T extends IAnyType> extends Array<Instance<T>> implement
     return Array;
   }
 
-  [$type]?: [this] | [any];
+  readonly [$env]: any;
+  readonly [$parent]: IStateTreeNode | null;
+  readonly [$type]: [this] | [any];
+
+  constructor(type: any, parent: IStateTreeNode | null, env: any) {
+    super();
+    this[$type] = type;
+    this[$parent] = parent;
+    this[$env] = env;
+  }
 
   get [Symbol.toStringTag]() {
     return "Array" as const;
@@ -67,24 +76,12 @@ class ArrayType<T extends IAnyType> extends BaseType<Array<T["InputType"]> | und
   }
 
   instantiate(snapshot: this["InputType"] | undefined, context: InstantiateContext, parent: IStateTreeNode | null): this["InstanceType"] {
-    const array = new QuickArray<T>(snapshot?.length ?? 0);
+    const array = new QuickArray<T>(this, parent, context.env);
     if (snapshot) {
-      for (let index = 0; index < array.length; ++index) {
-        array[index] = this.childrenType.instantiate(snapshot[index], context, array);
+      for (let index = 0; index < snapshot?.length; ++index) {
+        array.push(this.childrenType.instantiate(snapshot[index], context, array));
       }
     }
-
-    setParent(array, parent);
-
-    Reflect.defineProperty(array, $type, {
-      value: this,
-      configurable: false,
-      enumerable: false,
-      writable: false,
-    });
-
-    setEnv(array, context.env);
-
     return array as this["InstanceType"];
   }