… package.json
… pnpm-lock.yaml
… spec/array.spec.ts