Each commit shows the specifics, but to summarize:

### Don't use `for .. of` in some hot paths

https://v8.dev/blog/elements-kinds#avoid-reading-beyond-the-length-of-the-array

> Nowadays, the performance of both for-of and forEach is on par with the old-fashioned for loop.

But I don't observe this at all (that blog post is from 2017).

### More direct instantiation in our fast instantiator

* `types.frozen` can be directly assigned. There's technically a typeof check for functions, but I'm okay ignoring that for read-only roots (we can typecheck it out too, if desired)
* If an array's child type is directly assignable, the input snapshot will be exactly the format we need it.

### Simplify setting parent/env

We had some helpers to set the special parent/env values. These actually had to do a bunch of extra work when the caller had deeper knowledge about the objects being provided to these functions. I removed these functions in favour of all the different instantiators doing the necessary work.

One issue that arose from this change was that these properties are not always tagged as non-enumerable. This is okay forthings like `Object.keys` and `Object.entries`, which focus on the "own" string keys, but can now be enumerated by `getOwnPropertySymbols` and `getOwnPropertyDescriptors`. Jest appears to do something deeper for `toEqual` checks and these were getting pulled in. To circumvent this, I made array snapshots be proper arrays, not quick arrays.

✅ Speed
✅ Verified this doesn't break the main Gadget repo
✅ Tests passing

## Results

```
# Before
instantiating a small root x 1,698,169 ops/sec ±0.64% (98 runs sampled)
instantiating a large root x 4,168 ops/sec ±2.22% (95 runs sampled)
instantiating a large union x 326,144 ops/sec ±1.17% (95 runs sampled)

# After
instantiating a small root x 3,230,733 ops/sec ±0.28% (97 runs sampled)
instantiating a large root x 4,697 ops/sec ±1.57% (96 runs sampled)
instantiating a large union x 357,445 ops/sec ±1.36% (96 runs sampled)
```