@thegedge 2023-11-08
Gonna merge this one, and then let's get your PR in https://github.com/gadget-inc/mobx-quick-tree/pull/66. I'll cut a new release after.